WIP: add initial work in splitting up the Camp Map component into sizeable chunks (to then be able to fix performance)
